Pretty-up formating of model columns/relationships
This violates PEP8 in a couple ways:

* long lines
* extra space around/between tokens

I think this is much easier to read however.

This also reorders some columns, and I'm unclear if that matters as far
as the schema is concerned. I'm breaking the schema right now anyway
though, so I don't consider it a big deal. I'll worry about it when I
declare some level of stability.