Fix bug in assign_rounds
We weren't setting the time zone when pulling things out of the db, so
the equality comparison with various due dates wasn't working
correctly.

Python date/time libraries are garbage.

This also adds a unique constraint to the database which, while not
quite as strong as what we actually want to specify, should help catch
errors like this sooner.