Fix #37 and start refactoring show_status
Wow the view code is terrible. I essentially had to do some refactoring
just to understand what it was doing again. It's still not good enough
-- out to be split int separate functions, as it's still hard to follow
the big picture -- but it's better.

The new implementation also adds posts for which counts_for is None,
which is enough to make #37 work.