Split up tests/util
Randomized helpers and example data are now in their own files. Some of
the import statements in tests have been changes to accommodate this.

This also includes a fix to a bug introduced by
d6795017c49ccc0298395bf48c9b8452042e2aa5. It ought to be it's own
commit, but right now I just cannot be bothered. Basically: I
inadvertently removed the bit were we take the column out of our result,
so somewhere we ended up passing in a tuple where we wanted a datetime.