Fix post dedup bug.
We previously were not correctly handling the case where a post other
than the most recent was updated. I've simplified things a bit, so it no
longer tries to short-circut when it's hit an appropriately aged post,
and it looks for an older copy of each post regardless of where it
appears in the sequence.