Add types to ironblogger.date
This breaks compatibility with python 2 -- the types are worth it, and
I'm having trouble figuring out how to do the NewType stuff while using
the stub files exclusively.

Right now only ironblogger.date has been made type safe.

We've started stub files for the arrow library, which is one of our
dependencies that isn't in typeshed. We may upstream this at some point.